Wywalenie assert'ow bo to kurwa nie dziala

+++ src/main/java/pl/kamil0024/commands/moderation/PunishCommand.java
@@ -266,11 +266,12 @@ public class PunishCommand extends Command {
                 },
                 30, TimeUnit.SECONDS,
                 () -> {
-                    assert !kurwaBylaAkcja.get();
-                    try {
-                        msg.getChannel().sendMessage(String.format("<@%s>, twój czas na odpowiedź minął!", context.getUser().getId())).queue();
-                        msg.delete().complete();
-                    } catch (Exception ignored) { }
+                    if (!kurwaBylaAkcja.get()) {
+                        try {
+                            msg.getChannel().sendMessage(String.format("<@%s>, twój czas na odpowiedź minął!", context.getUser().getId())).queue();
+                            msg.delete().complete();
+                        } catch (Exception ignored) { }
+                    }
                 }
         );
     }
